Found the culprit for borked imports
The bug was that the .dyninst section was first written
and then the appropriate offsets were calculated
and the actuall PE section was created.
Along with that, the same VA was used both for disk offset and
memory offset, because , naturally , disk offset isn't known in
advance.

The mem offset calculated later when actualy building the PE section
is what's needed to calucalte the correct RVAs in writeImpTable.