Reverting the previous changes to emitWin and "fixing" the base issue.
In binaryEdit there is a piece of code there base of a new region is
adjusted and this is what seems to have been causing our problems
all along.
Commenting out those two lines fixes both imports and wrong
trampoline offsets (for future record, in dyninst trampolines are called
springboards).

Obviously, since this is in shared code, it probably breaks
something on linux or somewhere. Will ask on the list.