Changed function calling code emission
On linux, dyninst is using pc relative addressing
to call an instrumentation function and then relies
on elf relocs to resolve the symbol address at runtime.
We can't do that on windows, so instead, call directly into
iat entry. We don't know where iat is , so will have to fix it
later.

Similarly for external vars.